Split files and work on Lexer
- new VISType class to handle strings and type conversions

- Work on parsing

- Split string utils into its own file

+++ src/VISCommands.h
@@ -9,7 +9,7 @@ enum VISHeaderEnum {
   Data = 2,
 };
 
-typedef std::array<int, 3> VISWaypoint_t;
+typedef std::array<float, 3> VISWaypoint_t;
 
 
 class VISCommands {
